requirements.txt
beautifulsoup4>=4.12
camelot-py[cv]>=0.11
pandas>=2.0
requests>=2.31
supabase>=2.0
